Quick Settings fix for inversion and hotspot
Allow Quick Settings tiles to be arranged any way you want, delete them or add some.

Inversion and Hotspot tile are no longer contextual and can be added, removed or rearranged at will.

This is a partial pick of commit:
Add customization for Quick Settings [1/2]
https://github.com/CyanogenMod/android_frameworks_base/commit/e6f1d4e5caa8429c398bc922ac038a997780152e

Now in QS tuner the inversion and hotspot tiles are not there by default but can be added.
Changed the config to show the real default tiles and to add new tiles just add to "quick_settings_tiles_default"

Change-Id: Ie596a54a5c4eb777782fb047496cfcd0a6cd2e89