Bluetooth: Send Bluetooth Off intent based in BLEApp count
This change checks for the precense of an Always On
APP while moving Bluetooth state from ON to BLE_ON
state. If there is an app present, BT_OFF intent is
immidiately sent, else the intent is sent when BT is
completely turned OFF. This prevents the intermediate
state where BT Off intent, enables the BT toggle switch
while BT core services are still turning off.

Change-Id: I169af5a9aedd18e38ad990a793a437ac40cbe9ef